chore: move rule to `index.js`
* ensure consistency with other configured rules
